util.py::fullnameToGameDir: handle game name mismatches
Handle any discrepancies between the .csv collection cache
(which maps the full game name -> game directory name)
and the actual directory names in the collection config directory.
Should fix #119, though further testing may be required.

EXAMPLE:

The eXoDOSv6.csv collection cache maps the
full game name to the game directory name:
  * "Commander Keen - Keen Dreams (1993)" -> "CKeenDr"
But what if in a new release of eXoDOS, the game directory name were
changed to "ckeendr"? On a case-sensitive filesystem, the code looks for
"CKeenDr" (from the collection cache) and doesn't find it, but on a
case-insensitive filesystem there would be no issues (such as NTFS).
To avoid any issues, resolve the game directory from the cache to its
actual case-corrected filesystem name (in this case, "ckeendr")
before continuing.